Reorganize loadout drawer under a container component
