+++ src/app/loadout-drawer/loadout-drawer-reducer.ts
@@ -1,9 +1,9 @@
 import { LoadoutParameters } from '@destinyitemmanager/dim-api-types';
+import { D1ManifestDefinitions } from 'app/destiny1/d1-definitions';
+import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { t } from 'app/i18next-t';
 import { DimItem } from 'app/inventory/item-types';
-import { DimStore } from 'app/inventory/store-types';
 import { SocketOverrides } from 'app/inventory/store/override-sockets';
-import { getCurrentStore, getStore } from 'app/inventory/stores-helpers';
 import { showNotification } from 'app/notifications/notifications';
 import { itemCanBeInLoadout } from 'app/utils/item-utils';
 import { getSocketsByCategoryHash } from 'app/utils/socket-utils';
@@ -12,31 +12,12 @@ import { BucketHashes, SocketCategoryHashes } from 'data/d2/generated-enums';
 import produce from 'immer';
 import _ from 'lodash';
 import { Loadout, LoadoutItem, ResolvedLoadoutItem } from './loadout-types';
-import { newLoadout } from './loadout-utils';
 
 export interface State {
-  loadout?: Readonly<Loadout>;
-  /**
-   * The store that provides context to how this loadout is being edited from.
-   * The store this edit session was launched from. This is to help pick which
-   * mods are enabled, which subclass items to show, etc.
-   */
-  storeId?: string;
-  showClass: boolean;
-  isNew: boolean;
+  loadout: Readonly<Loadout>;
 }
 
 export type Action =
-  /** Reset the tool (for when the sheet is closed) */
-  | { type: 'reset' }
-  /** Start editing a new or existing loadout */
-  | {
-      type: 'editLoadout';
-      loadout: Loadout;
-      storeId: string;
-      isNew: boolean;
-      showClass: boolean;
-    }
   /** Replace the current loadout with an updated one */
   | { type: 'update'; loadout: Loadout }
   /** Add an item to the loadout */
@@ -46,7 +27,6 @@ export type Action =
       items: ResolvedLoadoutItem[];
       equip?: boolean;
       socketOverrides?: SocketOverrides;
-      stores: DimStore[];
     }
   /** Applies socket overrides to the supplied item */
   | { type: 'applySocketOverrides'; item: DimItem; socketOverrides: SocketOverrides }
@@ -62,43 +42,24 @@ export type Action =
 /**
  * All state for this component is managed through this reducer and the Actions above.
  */
-export function stateReducer(state: State, action: Action): State {
-  switch (action.type) {
-    case 'reset':
-      return {
-        showClass: true,
-        isNew: false,
-        loadout: undefined,
-      };
+export function stateReducer(_defs: D2ManifestDefinitions | D1ManifestDefinitions) {
+  return (state: State, action: Action): State => {
+    switch (action.type) {
+      case 'update':
+        return {
+          ...state,
+          loadout: action.loadout,
+        };
 
-    case 'editLoadout': {
-      const { loadout, storeId, isNew, showClass } = action;
+      case 'addItem': {
+        const { loadout } = state;
+        const { item, items, equip, socketOverrides } = action;
 
-      return {
-        ...state,
-        loadout,
-        storeId: storeId === 'vault' ? undefined : storeId,
-        isNew,
-        showClass,
-      };
-    }
+        if (!itemCanBeInLoadout(item)) {
+          showNotification({ type: 'warning', title: t('Loadouts.OnlyItems') });
+          return state;
+        }
 
-    case 'update':
-      return {
-        ...state,
-        loadout: action.loadout,
-      };
-
-    case 'addItem': {
-      const { loadout } = state;
-      const { item, items, equip, socketOverrides, stores } = action;
-
-      if (!itemCanBeInLoadout(item)) {
-        showNotification({ type: 'warning', title: t('Loadouts.OnlyItems') });
-        return state;
-      }
-
-      if (loadout) {
         if (item.classType !== DestinyClass.Unknown && loadout.classType !== item.classType) {
           showNotification({
             type: 'warning',
@@ -111,219 +72,184 @@ export function stateReducer(state: State, action: Action): State {
           ...state,
           loadout: draftLoadout,
         };
+      }
+
+      case 'removeItem': {
+        const { loadout } = state;
+        const { item, items } = action;
+        return loadout ? { ...state, loadout: removeItem(loadout, item, items) } : state;
+      }
+
+      case 'equipItem': {
+        const { loadout } = state;
+        const { item, items } = action;
+        return loadout ? { ...state, loadout: equipItem(loadout, item, items) } : state;
+      }
+
+      case 'applySocketOverrides': {
+        const { loadout } = state;
+        const { item, socketOverrides } = action;
+        return loadout
+          ? { ...state, loadout: applySocketOverrides(loadout, item, socketOverrides) }
+          : state;
+      }
+
+      case 'updateModsByBucket': {
+        const { loadout } = state;
+        const { modsByBucket } = action;
+        return loadout
+          ? {
+              ...state,
+              loadout: {
+                ...loadout,
+                parameters: {
+                  ...loadout.parameters,
+                  modsByBucket: _.isEmpty(modsByBucket) ? undefined : modsByBucket,
+                },
+              },
+            }
+          : state;
+      }
+
+      case 'updateMods': {
+        const { loadout } = state;
+        const { mods } = action;
+        return loadout
+          ? {
+              ...state,
+              loadout: {
+                ...loadout,
+                parameters: {
+                  ...loadout.parameters,
+                  mods,
+                },
+              },
+            }
+          : state;
+      }
+
+      case 'changeClearMods': {
+        const { loadout } = state;
+        const { enabled } = action;
+        return loadout
+          ? {
+              ...state,
+              loadout: {
+                ...loadout,
+                parameters: {
+                  ...loadout.parameters,
+                  clearMods: enabled,
+                },
+              },
+            }
+          : state;
+      }
+
+      case 'removeMod': {
+        const { loadout } = state;
+        const { hash } = action;
+        if (loadout) {
+          const newLoadout = { ...loadout };
+          const newMods = newLoadout.parameters?.mods?.length
+            ? [...newLoadout.parameters.mods]
+            : [];
+          const index = newMods.indexOf(hash);
+          if (index !== -1) {
+            newMods.splice(index, 1);
+            newLoadout.parameters = {
+              ...newLoadout.parameters,
+              mods: newMods,
+            };
+            return { ...state, loadout: newLoadout };
+          }
+        }
+        return state;
+      }
+    }
+  };
+}
+
+/**
+ * Produce a new loadout that adds a new item to the given loadout.
+ */
+function addItem(
+  loadout: Readonly<Loadout>,
+  item: DimItem,
+  items: ResolvedLoadoutItem[],
+  equip?: boolean,
+  socketOverrides?: SocketOverrides
+): Loadout {
+  const loadoutItem: LoadoutItem = {
+    id: item.id,
+    hash: item.hash,
+    amount: 1,
+    equip: false,
+  };
+
+  // TODO: maybe we should just switch back to storing loadout items in memory by bucket
+
+  // Other items of the same type (as DimItem)
+  const typeInventory = items.filter((li) => li.item.bucket.hash === item.bucket.hash);
+  const dupe = loadout.items.find((i) => i.hash === item.hash && i.id === item.id);
+  const maxSlots = item.bucket.capacity;
+
+  return produce(loadout, (draftLoadout) => {
+    const findItem = ({ loadoutItem }: ResolvedLoadoutItem) =>
+      draftLoadout.items.find((i) => i.id === loadoutItem.id && i.hash === loadoutItem.hash)!;
+
+    if (!dupe) {
+      if (typeInventory.length < maxSlots) {
+        loadoutItem.equip =
+          equip !== undefined ? equip : item.equipment && typeInventory.length === 0;
+        if (loadoutItem.equip) {
+          for (const otherItem of typeInventory) {
+            findItem(otherItem).equip = false;
+          }
+        }
+
+        // Only allow one subclass to be present per class (to allow for making a loadout that specifies a subclass for each class)
+        if (item.bucket.hash === BucketHashes.Subclass) {
+          const conflictingItem = items.find(
+            (li) => li.item.bucket.hash === item.bucket.hash && li.item.classType === item.classType
+          );
+          if (conflictingItem) {
+            draftLoadout.items = draftLoadout.items.filter((i) => i.id !== conflictingItem.item.id);
+          }
+          loadoutItem.equip = true;
+        }
+
+        if (socketOverrides) {
+          loadoutItem.socketOverrides = socketOverrides;
+        }
+
+        draftLoadout.items.push(loadoutItem);
+
+        // If adding a new armor item, remove any fashion mods (shader/ornament) that couldn't be slotted
+        if (
+          item.bucket.inArmor &&
+          loadoutItem.equip &&
+          draftLoadout.parameters?.modsByBucket?.[item.bucket.hash]?.length
+        ) {
+          const cosmeticSockets = getSocketsByCategoryHash(
+            item.sockets,
+            SocketCategoryHashes.ArmorCosmetics
+          );
+          draftLoadout.parameters.modsByBucket[item.bucket.hash] =
+            draftLoadout.parameters.modsByBucket[item.bucket.hash].filter((plugHash) =>
+              cosmeticSockets.some((s) => s.plugSet?.plugs.some((p) => p.plugDef.hash === plugHash))
+            );
+        }
       } else {
-        // If we don't have a loadout, this action was invoked via the "+ Loadout" button in item actions
-        let owner: DimStore =
-          item.owner === 'vault' ? getCurrentStore(stores)! : getStore(stores, item.owner)!;
-
-        if (item.classType !== DestinyClass.Unknown && item.classType !== owner.classType) {
-          const matchingStore = stores.find((s) => s.classType === item.classType);
-          if (!matchingStore) {
-            showNotification({
-              type: 'warning',
-              title: t('Loadouts.ClassTypeMissing', { className: item.classTypeNameLocalized }),
-            });
-            return state;
-          }
-          owner = matchingStore;
-        }
-
-        const classType =
-          item.classType === DestinyClass.Unknown ? owner.classType : item.classType;
-        const draftLoadout = addItem(
-          newLoadout('', [], classType),
-          item,
-          items,
-          equip,
-          socketOverrides
-        );
-        return {
-          ...state,
-          loadout: draftLoadout,
-          storeId: owner.id,
-          isNew: true,
-        };
+        showNotification({
+          type: 'warning',
+          title: t('Loadouts.MaxSlots', { slots: maxSlots }),
+        });
       }
+    } else if (item.maxStackSize > 1) {
+      const increment = Math.min(dupe.amount + item.amount, item.maxStackSize) - dupe.amount;
+      dupe.amount += increment;
     }
-
-    case 'removeItem': {
-      const { loadout } = state;
-      const { item, items } = action;
-      return loadout ? { ...state, loadout: removeItem(loadout, item, items) } : state;
-    }
-
-    case 'equipItem': {
-      const { loadout } = state;
-      const { item, items } = action;
-      return loadout ? { ...state, loadout: equipItem(loadout, item, items) } : state;
-    }
-
-    case 'applySocketOverrides': {
-      const { loadout } = state;
-      const { item, socketOverrides } = action;
-      return loadout
-        ? { ...state, loadout: applySocketOverrides(loadout, item, socketOverrides) }
-        : state;
-    }
-
-    case 'updateModsByBucket': {
-      const { loadout } = state;
-      const { modsByBucket } = action;
-      return loadout
-        ? {
-            ...state,
-            loadout: {
-              ...loadout,
-              parameters: {
-                ...loadout.parameters,
-                modsByBucket: _.isEmpty(modsByBucket) ? undefined : modsByBucket,
-              },
-            },
-          }
-        : state;
-    }
-
-    case 'updateMods': {
-      const { loadout } = state;
-      const { mods } = action;
-      return loadout
-        ? {
-            ...state,
-            loadout: {
-              ...loadout,
-              parameters: {
-                ...loadout.parameters,
-                mods,
-              },
-            },
-          }
-        : state;
-    }
-
-    case 'changeClearMods': {
-      const { loadout } = state;
-      const { enabled } = action;
-      return loadout
-        ? {
-            ...state,
-            loadout: {
-              ...loadout,
-              parameters: {
-                ...loadout.parameters,
-                clearMods: enabled,
-              },
-            },
-          }
-        : state;
-    }
-
-    case 'removeMod': {
-      const { loadout } = state;
-      const { hash } = action;
-      if (loadout) {
-        const newLoadout = { ...loadout };
-        const newMods = newLoadout.parameters?.mods?.length ? [...newLoadout.parameters.mods] : [];
-        const index = newMods.indexOf(hash);
-        if (index !== -1) {
-          newMods.splice(index, 1);
-          newLoadout.parameters = {
-            ...newLoadout.parameters,
-            mods: newMods,
-          };
-          return { ...state, loadout: newLoadout };
-        }
-      }
-      return state;
-    }
-  }
-
-  /**
-   * Produce a new loadout that adds a new item to the given loadout.
-   */
-  function addItem(
-    loadout: Readonly<Loadout>,
-    item: DimItem,
-    items: ResolvedLoadoutItem[],
-    equip?: boolean,
-    socketOverrides?: SocketOverrides
-  ): Loadout {
-    const loadoutItem: LoadoutItem = {
-      id: item.id,
-      hash: item.hash,
-      amount: 1,
-      equip: false,
-    };
-
-    // TODO: maybe we should just switch back to storing loadout items in memory by bucket
-
-    // Other items of the same type (as DimItem)
-    const typeInventory = items.filter((li) => li.item.bucket.hash === item.bucket.hash);
-    const dupe = loadout.items.find((i) => i.hash === item.hash && i.id === item.id);
-    const maxSlots = item.bucket.capacity;
-
-    return produce(loadout, (draftLoadout) => {
-      const findItem = ({ loadoutItem }: ResolvedLoadoutItem) =>
-        draftLoadout.items.find((i) => i.id === loadoutItem.id && i.hash === loadoutItem.hash)!;
-
-      if (!dupe) {
-        if (typeInventory.length < maxSlots) {
-          loadoutItem.equip =
-            equip !== undefined ? equip : item.equipment && typeInventory.length === 0;
-          if (loadoutItem.equip) {
-            for (const otherItem of typeInventory) {
-              findItem(otherItem).equip = false;
-            }
-          }
-
-          // Only allow one subclass to be present per class (to allow for making a loadout that specifies a subclass for each class)
-          if (item.bucket.hash === BucketHashes.Subclass) {
-            const conflictingItem = items.find(
-              (li) =>
-                li.item.bucket.hash === item.bucket.hash && li.item.classType === item.classType
-            );
-            if (conflictingItem) {
-              draftLoadout.items = draftLoadout.items.filter(
-                (i) => i.id !== conflictingItem.item.id
-              );
-            }
-            loadoutItem.equip = true;
-          }
-
-          if (socketOverrides) {
-            loadoutItem.socketOverrides = socketOverrides;
-          }
-
-          draftLoadout.items.push(loadoutItem);
-
-          // If adding a new armor item, remove any fashion mods (shader/ornament) that couldn't be slotted
-          if (
-            item.bucket.inArmor &&
-            loadoutItem.equip &&
-            draftLoadout.parameters?.modsByBucket?.[item.bucket.hash]?.length
-          ) {
-            const cosmeticSockets = getSocketsByCategoryHash(
-              item.sockets,
-              SocketCategoryHashes.ArmorCosmetics
-            );
-            draftLoadout.parameters.modsByBucket[item.bucket.hash] =
-              draftLoadout.parameters.modsByBucket[item.bucket.hash].filter((plugHash) =>
-                cosmeticSockets.some((s) =>
-                  s.plugSet?.plugs.some((p) => p.plugDef.hash === plugHash)
-                )
-              );
-          }
-        } else {
-          showNotification({
-            type: 'warning',
-            title: t('Loadouts.MaxSlots', { slots: maxSlots }),
-          });
-        }
-      } else if (item.maxStackSize > 1) {
-        const increment = Math.min(dupe.amount + item.amount, item.maxStackSize) - dupe.amount;
-        dupe.amount += increment;
-      }
-    });
-  }
+  });
 }
 
 /**
